### 2024-02-23 #6398 @editorialbot
[PRE REVIEW]: pyGCodeDecode: A Python package for time-accurate GCode simulation in material extrusion processes
**Submitting author:** @FelixFroelich (<a href="http://orcid.org/0000-0002-2575-4752">Felix Frölich</a>)
**Repository:** https://github.com/FAST-LB/pyGCodeDecode
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0
