### 2024-02-27 5c1f27a
Update paper.bib so DOIs are found
