+++ paper/paper.bib
@@ -36,17 +36,17 @@
 
 @book{Gibson:2021,
   Author = {{Gibson}, I. and {Rosen}, D.},
-  url = {https://doi.org/10.1007/978-3-030-56127-7},
+  doi = {10.1007/978-3-030-56127-7},
   Booktitle = {Additive Manufacturing Technologies ~ISBN 978-3-030-56126-0},
   Publisher = {Springer Nature Switzerland},
   Title = {{Additive Manufacturing Technologies: Third Edition}},
-  Year = {2021},
+  Year = 2021,
   page = {180}
 }
 
 
 @article{Yadav:2023,
-  url = {https://doi.org/10.1007/s11665-023-08476-2},
+  doi = {10.1007/s11665-023-08476-2},
   Author = {{Yadav}, K. and {Rohilla}, S and {Ali}, A},
   Journal = {Journal of Materials Engineering and Performance},
   Title = {{Effect of Speed, Acceleration, and Jerk on Surface Roughness of FDM-Fabricated Parts}},
@@ -54,7 +54,7 @@
 }
 
 @article{Luzanin:2019,
-  url = {http://dx.doi.org/10.1108/RPJ-12-2018-0316},
+  doi = {10.1108/RPJ-12-2018-0316},
   Author = {{Luzanin}, O. and {Movrin}, D.},
   Journal = {Rapid Prototyping},
   Title = {{Impact of processing parameters on tensile strength, in-process crystallinity and mesostructure in FDM-fabricated PLA specimens}},
@@ -62,7 +62,7 @@
 }
 
 @article{Altan:2018,
-  url = {https://doi.org/10.3139/120.111178},
+  doi = {doi.org/10.3139/120.111178},
   Author = {{Altan}, M. and {Eryildiz}, M.},
   Journal = {MATERIALS TESTING FOR JOINING AND ADDITIVE MANUFACTURING APPLICATIONS},
   Title = {{Effects of process parameters on the quality of PLA products fabricated by fused deposition modeling (FDM): surface roughness and tensile strength}},
