### 2024-02-27 #6398 @lukashof
We changed the bib so the DOIs should be picked up now.
