### 2024-02-28 #6398 @FelixFroelich
> @lukashof Do you have any suggested reviewers?

In the engineering field, we know Nicolas Christ [Extraweich](https://github.com/Extraweich/Extraweich) . He has also published something in the JOSS Journal (https://github.com/Extraweich/homopy) and we think he would be suitable to review this project.
