### 2024-02-28 30e6d42
spelling and some return hints
