+++ doc.md
@@ -25,7 +25,7 @@ Generate abaqus event series.
 - `tolerance` _float, default = 1e-12_ - tolerance to determine whether extrusion is happening
 - `output_unit_system` _str, optional_ - Unit system for the output.
   The one from the simulation is used, in None is specified.
-  
+
 
 **Returns**:
 
@@ -67,7 +67,7 @@ Convert list of states to trajectory repr. by planner blocks.
 
 - `states` - (list[state]) list of states
 - `firmware` - (string, default = None) select firmware by name
-  
+
 
 **Returns**:
 
@@ -92,7 +92,7 @@ Find the current segment.
 - `t` - (float) time of search
 - `last_index` - (int) last found index for optimizing search
 - `keep_position` - (bool) keeps position of last segment, use this when working with gaps of no movement between segments
-  
+
 
 **Returns**:
 
@@ -112,7 +112,7 @@ Return list of segments by unpacking list of planner blocks.
 **Arguments**:
 
 - `blocklist` - (list[planner_block]) list of planner blocks
-  
+
 
 **Returns**:
 
@@ -151,7 +151,7 @@ Initialize the Simulation of a given G-code with initial machine setup or defaul
   machine name: (string, default = None) name of the default machine to use
 - `initial_machine_setup` - (setup, default = None) setup instance
 - `output_unit_system` - (string, default = "SImm") available unit systems: SI, SImm & inch
-  
+
 
 **Example**:
 
@@ -216,7 +216,7 @@ Plot axis velocity with matplotlib.
 - `time_steps` - (int or string, default = "constrained") number of time steps or constrain plot vertices to segment vertices
 - `filepath` - (Path, default = None) save fig as image if filepath is provided
 - `dpi` - (int, default = 400) select dpi
-  
+
 
 **Returns**:
 
@@ -248,7 +248,7 @@ Return unit system scaled values for vel and pos.
 - `t` - (float) time
 - `output_unit_system` _str, optional_ - Unit system for the output.
   The one from the simulation is used, in None is specified.
-  
+
 
 **Returns**:
 
@@ -311,12 +311,12 @@ Return scaled xyz min & max while extruding.
 
 - `output_unit_system` _str, optional_ - Unit system for the output.
   The one from the simulation is used, in None is specified.
-  
+
 
 **Raises**:
 
 - `ValueError` - if nothing is extruded
-  
+
 
 **Returns**:
 
@@ -336,7 +336,7 @@ Return scaled maximum velocity while extruding.
 
 - `output_unit_system` _str, optional_ - Unit system for the output.
   The one from the simulation is used, in None is specified.
-  
+
 
 **Returns**:
 
@@ -355,7 +355,7 @@ Save summary to .yaml file.
 **Arguments**:
 
 - `filepath` _pathlib.Path | str_ - path to summary file
-  
+
   Saved data keys:
   - filename (string, filename)
   - t_end (float, end time)
@@ -376,7 +376,7 @@ Get a scaling factor to convert lengths from mm to another supported unit system
 
 - `output_unit_system` _str, optional_ - Wanted output unit system.
   Uses the one specified for the simulation on None is specified.
-  
+
 
 **Returns**:
 
@@ -454,7 +454,7 @@ Set initial Position.
 - `initial_position` - (tuple or dict) set initial position as tuple of len(4) or dictionary with keys: {X, Y, Z, E}.
[…]
