### 2024-02-29 03feaf5
fixing type hint to work with python 3.8
