+++ pyGCodeDecode/utils.py
@@ -8,6 +8,8 @@ Utils for the GCode Reader contains:
     - position
 """
 
+from typing import List
+
 import numpy as np
 
 
@@ -164,7 +166,7 @@ class vector_4D:
             ):
                 return True
 
-    def get_vec(self, withExtrusion=False) -> list[float]:
+    def get_vec(self, withExtrusion=False) -> List[float]:
         """Return the 4D vector, optionally with extrusion.
 
         Args:
@@ -191,7 +193,7 @@ class vector_4D:
 
 
 class velocity(vector_4D):
-    """4D - Velocity object for (cartesian) 3D printer."""
+    """4D - Velocity object for (Cartesian) 3D printer."""
 
     def __str__(self) -> str:
         """Print out velocity."""
@@ -249,7 +251,7 @@ class velocity(vector_4D):
 
 
 class position(vector_4D):
-    """4D - Position object for (cartesian) 3D printer."""
+    """4D - Position object for (Cartesian) 3D printer."""
 
     def __str__(self) -> str:
         """Print out position."""
@@ -279,9 +281,9 @@ class position(vector_4D):
         Returns:
             is_extruding: (bool) true if between self and other is extrusion
         """
-        extr = other.e - self.e if ignore_retract else abs(other.e - self.e)
+        extrusion = other.e - self.e if ignore_retract else abs(other.e - self.e)
 
-        if extr > 0:
+        if extrusion > 0:
             return True
         else:
             return False
