### 2024-03-01 8af664f
fixing the type hints for ´extrusion_max_vel()´ and ´extrusion_block_max_vel()´
