+++ paper/paper.bib
@@ -62,7 +62,7 @@
 }
 
 @article{Altan:2018,
-  doi = {doi.org/10.3139/120.111178},
+  doi = {10.3139/120.111178},
   Author = {{Altan}, M. and {Eryildiz}, M.},
   Journal = {MATERIALS TESTING FOR JOINING AND ADDITIVE MANUFACTURING APPLICATIONS},
   Title = {{Effects of process parameters on the quality of PLA products fabricated by fused deposition modeling (FDM): surface roughness and tensile strength}},
