+++ .github/workflows/tests.yaml
@@ -26,7 +26,7 @@ jobs:
         strategy:
             matrix:
                 os: [ubuntu-latest] # [ubuntu-latest, windows-latest, macos-latest]
-                python-version: ["3.10"] # ["3.8", "3.9", "3.10", "3.11", "3.12"]
+                python-version: ["3.9", "3.10", "3.11", "3.12"]
         steps:
           - name: checkout
             uses: actions/checkout@v4
