### 2024-04-04 #6465 @Extraweich
Hi everybody,
Sorry for being absent for a while. I was busy with conferences and my own publications, but I will take my time in the following week to hopefully finish the review.

As a starting point I would like to highlight some issues in the paper:

- Line 49: it would be preferable to add some references or links for the given documentations/algorithms (i.e. MARLIN classic jerk, MARLIN junction deviation and Klipper). Not being part of the 3D printing research community, I would want something to read into.
- Figure 2 and following equations: the figure is inconsistent in the sense that you are using $t_{acc}$, $t_{const}$ and $t_{dec}$ as time increments or intervals, but in the current notation they can easily be interpreted as specific locations in time. As such, it might add clearity if you would indicate their incremental nature, e.g. $\Delta t_{acc}$, $\Delta t_{const}$ and $\Delta t_{dec}$, as it is also indicated in the figure itself with $a=\frac{\Delta v}{\Delta t}$. Also the font differs between figure and equations, since you are using italic subscripts in the figure, but not in the equations.
- Line 67: what is indicted by "..." following GRBL/MARLIN?

In general, I am missing a comparison with other softwares which are out there and highlighting in what sense pyGCodeDecode is different (state of the field). Please add a section "Statement of Need" for this, which is required by JOSS.

Best regards
Nicolas
