+++ paper/paper.md
@@ -52,9 +52,14 @@ Setting a higher target printing velocity on a machine with insufficient acceler
 
 ![Printing velocity of the raw GCode (left) in comparison to the printing velocity with simulated acceleration (right). \label{fig:acc_comp}](comparison.png){width=80%}
 
+# Statement of need
+
+There are several tools and software available to view the contents of a GCode file, for example, integrated into the slicer software such as Prusa Slicer or Cura. But also available through web applications and printer controller applications like Octoprint and Repetier-Host. These tools are able to read the position of the GCode coordinates and interpolate between the points, it is possible to differentiate between print moves and travel moves. The additional information contained in the GCode such as target printing velocity or temperature can also be displayed.
+Without the interpretation and segment modeling that <span style="font-variant:small-caps;">pyGCodeDecode</span> applies, the mentioned software is not able to accurately model the printer's behavior, including the acceleration and deceleration segments of the movement as well as its influence on neighboring segments. This can result in inaccurate time predictions as well as possibly unrecognized deviations from the expected process parameters. To adress this issue, <span style="font-variant:small-caps;">pyGCodeDecode</span> was developed as an open-source firmware simulation tool. It enables more detailed and accurate simulation models for FFF-based processes by taking firmware behavior into account.
+
 # Methodology
 
-<span style="font-variant:small-caps;">pyGCodeDecode's</span> class-based structure and separation of modules allow for simple and extensive modifications or additions. Its GCode parser transfers individual commands into a state class containing every command's parameters as well as the GCode history and user-set firmware parameters. Most printers use a trapezoidal velocity profile for each move which is constrained by its entry, target and exit velocities, as well as the maximum acceleration. While the maximum acceleration and target velocity are configured in the firmware settings and the GCode respectively, the entry and exit velocities are calculated using a variety of different cornering algorithms. Usually some limited instantaneous change in velocity is allowed, while taking the change in travel direction into account. Smaller changes in direction generally require less reduction in travel speed. <span style="font-variant:small-caps;">pyGCodeDecode</span> provides models of cornering algorithms for several firmwares. They are implemented as classes according to the respective documentation, e.g. <span style="font-variant:small-caps;">Marlin</span> classic jerk, <span style="font-variant:small-caps;">Marlin</span> junction deviation and Klipper. The junction velocities are calculated using the selected cornering algorithm. Then the trajectory modeling connects all states by planning accelerating, constant velocity, and decelerating segments matching the junction velocities. This is achieved by solving the equations of the surface area under the trapezoidal velocity profile shown in \autoref{fig:trapezoid} for the missing parameters.
[…]
