### 2024-04-05 1084d2a
adding .cff, CoC, Contribution guidlines and "MEX" as keyword
