+++ paper/paper.md
@@ -13,6 +13,7 @@ tags:
   - Firmware
   - FFF
   - FDM
+  - MEX
   - Process Simulation
 authors:
   - name: Jonathan Knirsch
