+++ CITATION.cff
@@ -1,0 +1,63 @@
+# This CITATION.cff file was generated with cffinit.
+# Visit https://bit.ly/cffinit to generate yours today!
+
+cff-version: 1.2.0
+title: pyGCodeDecode
+message: >-
+  You are welcome to cite pyGCodeDecode using the metadata
+  from this file.
+type: software
+authors:
+  - name: FAST-LB at the Karlsruhe Institute of Technology (KIT)
+    address: Rintheimer Queeralle 2
+    city: Karlsruhe
+    post-code: '76131'
+    location: Building 70.04
+    email: lt-github@fast.kit.edu
+    website: 'https://www.fast.kit.edu/english/lbt/index.php'
+    country: DE
+  - given-names: Jonathan
+    family-names: Knirsch
+    email: jonathan.knirsch@student.kit.edu
+    affiliation: FAST-LB at KIT
+    orcid: 'https://orcid.org/0009-0008-8663-8721'
+  - given-names: Felix
+    family-names: Frölich
+    email: felix.froelich@kit.edu
+    affiliation: FAST-LB at KIT
+    orcid: 'https://orcid.org/0000-0002-2575-4752'
+  - given-names: Lukas
+    family-names: Hof
+    orcid: 'https://orcid.org/0009-0000-4931-2409'
+    affiliation: FAST-LB at KIT
+    email: lukas.hof@kit.edu
+  - given-names: Florian
+    family-names: Wittemann
+    email: florian.wittemann@kit.edu
+    affiliation: FAST-LB at KIT
+    orcid: 'https://orcid.org/0000-0001-8693-7264'
+  - given-names: Luise
+    family-names: Kärger
+    email: luise.kaerger@kit.edu
+    affiliation: FAST-LB at KIT
+    orcid: 'https://orcid.org/0000-0002-6534-6518'
+repository-code: 'https://github.com/FAST-LB/pyGCodeDecode'
+abstract: >-
+  Generate time dependent boundary conditions from a .gcode
+  file.
+keywords:
+  - GCode
+  - Simulation
+  - Fused Filament Deposition
+  - Fused Filament Fabrication
+  - Material Extrusion
+  - Additive Manufacturing
+  - 3D Printing
+  - Machine Code
+  - Firmware
+  - FFF
+  - FDM
+  - MEX
+  - Process Simulation
+  - Engineering
+license: MIT
