+++ paper/paper.md
@@ -54,8 +54,8 @@ Setting a higher target printing velocity on a machine with insufficient acceler
 
 # Statement of need
 
-There are several tools and software available to view the contents of a GCode file, for example, integrated into the slicer software such as Prusa Slicer or Cura. But also available through web applications and printer controller applications like Octoprint and Repetier-Host. These tools are able to read the position of the GCode coordinates and interpolate between the points, it is possible to differentiate between print moves and travel moves. The additional information contained in the GCode such as target printing velocity or temperature can also be displayed.
-Without the interpretation and segment modeling that <span style="font-variant:small-caps;">pyGCodeDecode</span> applies, the mentioned software is not able to accurately model the printer's behavior, including the acceleration and deceleration segments of the movement as well as its influence on neighboring segments. This can result in inaccurate time predictions as well as possibly unrecognized deviations from the expected process parameters. To adress this issue, <span style="font-variant:small-caps;">pyGCodeDecode</span> was developed as an open-source firmware simulation tool. It enables more detailed and accurate simulation models for FFF-based processes by taking firmware behavior into account.
+There are several tools and software to visualize the GCode file data. For example, in various slicer programs such as Prusa Slicer or Cura, but also in web applications and printer control applications such as Octoprint and Repetier-Host. These tools can read the position of the GCode coordinates and interpolate between the points to distinguish between printing and traversing motions. The additional information in the GCode, such as target print speed or temperature, can also be displayed in most cases. However, currently available tools are unable to accurately simulate the behavior of the printer, including the acceleration and deceleration segments of the motion and their influence on adjacent segments. This can lead to inaccurate time predictions and potentially undetected deviations from expected process conditions. 
+The variety of tools available to visualize and interpret the G-Code underscores the importance of such data simulations. In addition, the constant and rapid advancement of printing technologies requires a deeper understanding of printer-specific process conditions, which must take into account hardware and firmware limitations. To fill this gap, <span style="font-variant:small-caps;">pyGCodeDecode</span> has been developed as an open source firmware simulation tool. It enables more detailed and accurate simulation models for MEX-based processes by taking into account the behavior of the firmware.
 
 # Methodology
 
