### 2024-04-08 a707f4a
Write company names with small capitals.
