+++ paper/paper.md
@@ -54,7 +54,7 @@ Setting a higher target printing velocity on a machine with insufficient acceler
 
 # Statement of need
 
-There are several software to visualize GCode file data. For example, in various slicer programs such as <span style="font-variant:small-caps;">Prusa Slicer</span> or <span style="font-variant:small-caps;">Cura</span>, but also in web applications and printer control applications such as <span style="font-variant:small-caps;">Octoprint</span>, <span style="font-variant:small-caps;">Repetier-Host</span>, <span style="font-variant:small-caps;">NC Viewer</span> or <span style="font-variant:small-caps;">gCode Viewer</span>. These tools can read the position of the GCode coordinates and interpolate between the points to create motion paths. It is possible to distinguish between printing and traversing motions in order to preview the part. The additional information in the GCode, such as target print speed or temperature, can also be displayed in most cases. However, currently available tools are unable to accurately simulate the behavior of the printer, including the acceleration and deceleration segments of the motion and their influence on adjacent segments. This can lead to inaccurate time predictions and potentially undetected deviations from expected process conditions.
+There are several software to visualize GCode file data. For example, in various slicer programs such as <span style="font-variant:small-caps;">Prusa Slicer</span> [@prusa_slicer] or <span style="font-variant:small-caps;">Cura</span> [@cura], but also in web applications and printer control applications such as <span style="font-variant:small-caps;">Octoprint</span> [@octoprint], <span style="font-variant:small-caps;">Repetier-Host</span> [@repetier_host], <span style="font-variant:small-caps;">NC Viewer</span> [@ncviewer] or <span style="font-variant:small-caps;">gCode Viewer</span> [@gcodeviewer]. These tools can read the position of the GCode coordinates and interpolate between the points to create motion paths. It is possible to distinguish between printing and traversing motions in order to preview the part. The additional information in the GCode, such as target print speed or temperature, can also be displayed in most cases. However, currently available tools are unable to accurately simulate the behavior of the printer, including the acceleration and deceleration segments of the motion and their influence on adjacent segments. This can lead to inaccurate time predictions and potentially undetected deviations from expected process conditions.
 The variety of software tools available underscores the importance of being able to analyze the GCode. In addition, the constant and rapid advancement of printing technologies requires a deeper understanding of printer-specific process conditions, which must take into account hardware and firmware limitations. To fill this gap, <span style="font-variant:small-caps;">pyGCodeDecode</span> has been developed as an open source firmware simulation tool. It enables more detailed and accurate simulation models for MEX-based processes by taking into account the behavior of the firmware.
 
 # Methodology
