+++ paper/paper.bib
@@ -68,3 +68,51 @@
   Title = {{Effects of process parameters on the quality of PLA products fabricated by fused deposition modeling (FDM): surface roughness and tensile strength}},
   Year = 2018
 }
+
+@software{cura,
+  author = {{Ultimaker B.V.}},
+  license = {LGPL-3.0},
+  month = jul,
+  title = {{Ultimaker Cura}},
+  url = {https://github.com/Ultimaker/Cura},
+  version = {5.4.0},
+  year = {2023}
+}
+
+@software{prusa_slicer,
+  author = {{Prusa Research a.s.}},
+  license = {AGPL-3.0},
+  title = {{PrusaSlicer}},
+  url = {https://github.com/prusa3d/PrusaSlicer},
+  version = {2.7.4},
+  year = {2024}
+}
+
+@software{octoprint,
+  author = {{Gina Häußge}},
+  license = {AGPL-3.0},
+  title = {{OctoPrint}},
+  url = {https://github.com/OctoPrint/OctoPrint},
+  version = {1.9.3},
+  year = {2023}
+}
+
+@software{ncviewer,
+  author = {{Xander Luciano}},
+  title = {{NC Viewer}},
+  url = {https://ncviewer.com/},
+  version = {1.1.3}
+}
+
+@software{gcodeviewer,
+  author = {{Alex Ustyantsev}},
+  title = {{gCodeViewer}},
+  url = {https://gcode.ws/}
+}
+
+@software{repetier_host,
+  author = {{Hot-World GmbH & Co. KG}},
+  title = {{Repetier-Host}},
+  url = {https://www.repetier.com/},
+  version = {2.3.2}
+}
