+++ paper/paper.bib
@@ -1,35 +1,30 @@
-@misc{grbl,
+@software{grbl,
   author = {Sungeun K. Jeon},
   title = {GRBL Firmware},
   year = {2021},
-  publisher = {},
-  journal = {GitHub repository},
+  version = {1.1}
   url = {https://github.com/grbl/grbl}
 }
 
-@misc{marlin,
+@software{marlin,
   author = {Scott Lahteine},
   title = {Marlin Firmware},
-  year = {2023},
-  publisher = {},
-  journal = {GitHub repository},
+  version = {2.1.2.2}
   url = {https://github.com/MarlinFirmware/Marlin}
 }
 
 @misc{marlin_doc,
   author = {Scott Lahteine},
   title = {Marlin Documentation},
-  year = {2023},
-  publisher = {},
+  note = {{Accessed: April 9\textsuperscript{th} 2024}},
   journal = {},
   url = {https://marlinfw.org/}
 }
 
-@book{klipper_doc,
+@misc{klipper_doc,
   author = {Klipper3d},
   title = {Klipper Documentation},
-  year = {2023},
-  publisher = {},
+  note = {{Accessed: April 9\textsuperscript{th} 2024}},
   journal = {},
   url = {https://www.klipper3d.org/Kinematics.html}
 }
@@ -100,6 +95,7 @@
 @software{ncviewer,
   author = {{Xander Luciano}},
   title = {{NC Viewer}},
+  note = {{Accessed: April 9\textsuperscript{th} 2024}},
   url = {https://ncviewer.com/},
   version = {1.1.3}
 }
@@ -107,6 +103,7 @@
 @software{gcodeviewer,
   author = {{Alex Ustyantsev}},
   title = {{gCodeViewer}},
+  note = {{Accessed: April 9\textsuperscript{th} 2024}},
   url = {https://gcode.ws/}
 }
 
