+++ paper/paper.bib
@@ -2,14 +2,14 @@
   author = {Sungeun K. Jeon},
   title = {GRBL Firmware},
   year = {2021},
-  version = {1.1}
+  version = {1.1},
   url = {https://github.com/grbl/grbl}
 }
 
 @software{marlin,
   author = {Scott Lahteine},
   title = {Marlin Firmware},
-  version = {2.1.2.2}
+  version = {2.1.2.2},
   url = {https://github.com/MarlinFirmware/Marlin}
 }
 
