+++ paper/paper.bib
@@ -16,7 +16,7 @@
 @misc{marlin_doc,
   author = {Scott Lahteine},
   title = {Marlin Documentation},
-  note = {{Accessed: April 9\textsuperscript{th} 2024}},
+  urldate = {2024-04-09},
   journal = {},
   url = {https://marlinfw.org/}
 }
@@ -24,7 +24,7 @@
 @misc{klipper_doc,
   author = {Klipper3d},
   title = {Klipper Documentation},
-  note = {{Accessed: April 9\textsuperscript{th} 2024}},
+  urldate = {2024-04-09},
   journal = {},
   url = {https://www.klipper3d.org/Kinematics.html}
 }
@@ -95,15 +95,15 @@
 @software{ncviewer,
   author = {{Xander Luciano}},
   title = {{NC Viewer}},
-  note = {{Accessed: April 9\textsuperscript{th} 2024}},
   url = {https://ncviewer.com/},
+  urldate = {2024-04-09},
   version = {1.1.3}
 }
 
 @software{gcodeviewer,
   author = {{Alex Ustyantsev}},
   title = {{gCodeViewer}},
-  note = {{Accessed: April 9\textsuperscript{th} 2024}},
+  urldate = {2024-04-09},
   url = {https://gcode.ws/}
 }
 
