### 2024-04-09 #6465 @FelixFroelich
> > Hi everybody, Sorry for being absent for a while. I was busy with conferences and my own publications, but I will take my time in the following week to hopefully finish the review.
> > As a starting point I would like to highlight some issues in the paper:
> >
> > * Line 49: it would be preferable to add some references or links for the given documentations/algorithms (i.e. MARLIN classic jerk, MARLIN junction deviation and Klipper). Not being part of the 3D printing research community, I would want something to read into.
> > * Figure 2 and following equations: the figure is inconsistent in the sense that you are using tacc, tconst and tdec as time increments or intervals, but in the current notation they can easily be interpreted as specific locations in time. As such, it might add clearity if you would indicate their incremental nature, e.g. Δtacc, Δtconst and Δtdec, as it is also indicated in the figure itself with a=ΔvΔt. Also the font differs between figure and equations, since you are using italic subscripts in the figure, but not in the equations.
> > * Line 67: what is indicted by "..." following GRBL/MARLIN?
> >
> > In general, I am missing a comparison with other softwares which are out there and highlighting in what sense pyGCodeDecode is different (state of the field). Please add a section "Statement of Need" for this, which is required by JOSS.
> > Concerning the check point "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support", it would be good to add a few more words in your README.md to clarify how exactly you expect people from the community to contribute to the project, or you could add a CONTRIBUTING.md (e.g. the one in my own [repository](https://github.com/Extraweich/homopy/blob/main/CONTRIBUTING.md) as a reference).
> > Best regards Nicolas
>
> Hi Nicolas,
>
> Thank you very much for reviewing our paper.
>
> We agree with all of your points and will incorporate them into the paper. As soon as the revised paper is in the main branch, I will let you know.
>
> We will also create a CONTRIBUTING.md file based on your repository and add the missing information.
>
> Best regards Felix

Hi everyone,

We have now revised the paper according to @Extraweich  comments.
We also added a CONTRIBUTING.md file and a Code_of_Conduct.md file to the repository.

All revisions have already been merged into the main branch. We hope all the changes are clear and understandable. If not, please do not hesitate to contact us with any questions!

Thanks for your helpful comments!

Best regards
Felix
