### 2024-05-26 #6465 @RahulSundar
Some preliminary comments are as follows:
1. Please do add some key references in the main README.md file as well of your repository.
2. Also, request that any set of step-by instructions are presented in the README.md as an enumerated list and utilize appropriate grammar, punctuation and capitalizations in the instructions where necessary.
3. Although a statement of need is provided where it is mentioned the other software are inaccurate, it would be useful to see an image comparing them in the validation section. This would further strengthen the claim of the paper.
4. In addition to showing a comparison, the errors need to be tabulated as well.
