### 2024-05-26 #6465 @RahulSundar
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/9242572046?check_suite_focus=true).

@zhubonan why is this happening?
