### 2024-05-27 #6465 @lukashof
Dear @RahulSundar,
thank you for your remarks! We will look at what we can do to address those as soon as possible and come back to you. We are looking forward to your further input! 🙂

Dear @zhubonan,
we don't understand why the PDF-generation is failing. The error message states that the repo couldn't be downloaded, but we didn't change the URL. Was there some recent change to the bot?
