### 2024-05-27 #6465 @zhubonan
Have you tried to generate the pdf (the simplest (the easiest way is to use docker)? If you can do that, feel free to attach it to this issue as a temporary workaround.
