### 2024-05-27 #6465 @lukashof
Yes, we compile a draft with every commit to main in our workflow. I'm attaching the most recent version.
[pyGCD_paper.pdf](https://github.com/openjournals/joss-reviews/files/15457405/pyGCD_paper.pdf)
