### 2024-05-28 #6465 @danielskatz
This seems to be working now, so perhaps it was a transient error.

Also, when one of us cloned the repo locally, we got this warning message:
```
Encountered 3 file(s) that should have been pointers, but weren't:
	tests/data/test_gcode_interpreter.gcode
	tests/data/test_state_generator.gcode
	tests/self_test/self_test.gcode
```
