### 2024-06-03 26c5abe
fix benchy example on Ubuntu
