### 2024-06-06 8b991ea
more introduction & namin max error
