+++ README.md
@@ -12,11 +12,11 @@
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
 
 ## What is this repository for?
+The analysis of GCode either in Slicer softwares or in dedicated GCode analyzer tools just shows target process parameters. The real process velocity is determined by the firmware of the real printer. The real velocity is expected to deviate from the target velocity at times of acceleration and deceleration. To model this behaviour `pyGCodeDecode` has been developed, which in contrast to existing software such as [Prusa Slicer][prusa_slicer] or [gCodeViewer][gcodeviewer] consideres and visualizes these influences. A comparison between the simulated acceleration approach and the raw GCode target velocity is illustrated below.
+
 ![comparison](https://media.githubusercontent.com/media/FAST-LB/pyGCodeDecode/main/paper/comparison.png)
 
-This package reads the target trajectory and commands for changing firmware settings from a GCode file. Furthermore it simulates a motion planner with acceleration and jerk / junction control. The simulation result describes the nozzle and extrusion axis position and velocity at every point in time. Notably, this method does try to simulate the real printer movements at a higher accuracy than assuming constant velocity. A comparison between the simulated accceleration approach and a constant velocity assumption is illustrated above. Constant velocity is usually also pictured in slicer previews and other GCode analyzing tools such as <span style="font-variant:small-caps;">Prusa Slicer</span> [1][prusa_slicer] or <span style="font-variant:small-caps;">gCodeViewer</span> [2][gcodeviewer]. The more accurate modeling is only achieved by replicating grbl and derivative firmwares specific movement planner solutions, such as Junction Deviation as an interpretation for Jerk. This python package can be used to generate time dependent boundary conditions from a GCode file, needed in additive manufacturing simulations such as Fused Filament Fabrication. With implemented 3D plotting functions, it also can be useful as a GCode analyzer tool, to visualize local velocities to gain better process understanding.
-
-The package is modularized to enable quick modification and extension of all features.
+In detail, this package reads the target trajectory and commands for changing firmware settings from a GCode file. Subsequently it simulates a motion planner with acceleration and jerk / junction control. The more accurate modeling can only be achieved by replicating grbl and derivative firmwares specific movement planner solutions, such as Junction Deviation as an interpretation for Jerk. The simulation result describes the nozzle and extrusion axis position and velocity at every point in time and are easily accessible. This python package can be used to generate time dependent boundary conditions from a GCode file, needed in additive manufacturing simulations such as Fused Filament Fabrication. With implemented 3D plotting functions, it also can be useful as a GCode analyzer tool, to visualize local velocities to gain better process understanding.
 
 
 ## Install pyGCodeDecode
@@ -75,7 +75,7 @@ For example, the definition may look like this (also see [./pyGCodeDecode/data/d
                 vZ: 12
                 vE: 80
                 firmware: marlin_jerk
-The default settings usually are machine specific and often can be read from the printer using a Serial Connection with a GCode command. you can use `M503` for Marlin, Prusa and some other firmwares.
+The default settings usually are machine specific and often can be read from the printer using a Serial Connection with a GCode command. You can use `M503` for Marlin, Prusa and some other firmwares.
 
 ### Use pyGCD to run a GCode Simulation
 
@@ -152,8 +152,9 @@ Only partially supported commands:
 Known unsupported commands that may cause issues:
 
         "G2" / "G3: {-} Arc/Circle move, please disable this command in your Slicer settings
-        
 
+
+ <!-- REFERENCES   -->
 [prusa_slicer]: <https://github.com/prusa3d/PrusaSlicer> "Prusa Slicer"
 
 [gcodeviewer]: <https://gcode.ws/> "gCodeViewer"
