### 2024-06-10 4fa6d91
catching errors when no display is available
