### 2024-06-11 b253677
small cleanup of test folder
