+++ tests/self_test/self_test.py
@@ -1,17 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Test pyGCD with known analytical solutions for simple trajectory planning tasks with other features as well."""
-
-
-# travel distance G0 / G1
-# extrusion length
-# time for simple movements
-# self correction check
-#
-# adhere to max acc
-# adhere to max vel
-# jerk / JD
-# units
-# layer detection / comment functionality
-# rel / abs pos/extr
-# set position
-#
