+++ tests/data/test_printer_setups.yaml
@@ -15,6 +15,21 @@ prusa_mini:
   vE: 80
   firmware: marlin_jd
 
+test:
+  # general properties
+  nozzle_diam: 0.4
+  filament_diam: 1.75
+  # default settings
+  p_vel: 35
+  p_acc: 1000
+  jerk: 10
+  # axis max speeds
+  vX: 180
+  vY: 180
+  vZ: 30
+  vE: 33
+  firmware: marlin_jd
+
 debugging:
   # general properties
   nozzle_diam: 0.4
