+++ tests/self_test/known_results.yaml
@@ -1,13 +1,0 @@
-# travel distance G0 / G1
-# extrusion length
-# time for simple movements
-# self correction check
-#
-# adhere to max acc
-# adhere to max vel
-# jerk / JD
-# units
-# layer detection / comment functionality
-# rel / abs pos/extr
-# set position
-#
