### 2024-06-11 8ed3d1f
adding rudimentary end to end tests
