### 2024-06-11 61ea958
adding missing pyvista dependencies in tests
