### 2024-06-11 0139653
adding 2d plot to end to end test
