+++ paper/paper.md
@@ -91,7 +91,7 @@ All segments of a single move are stored together with its enclosing states in a
 
 ![Validation of the simulation by measuring layer duration.\label{fig:validation}](plot.svg){width=60%}
 
-For the chosen case the layer duration is highly dependant on the set jerk values. For jerk values equal to the target printing velocity, the calculated time is expected to approach a constant velocity solution calculated analytically. Therefore, the acceleration and cornering algorithms have no influence on the print time of a layer. This case also represents the current state of the art, in which no further modeling is done. For jerk values close to zero, the printer is expected to slow almost to a full stop for each turn in the path. This result is similar to the simplest velocity trapezoid where entry and exit velocities are zero. The layer time for this edge case was also  validated by an analytical calculation. The comparison to experimental data for jerk values between these edge cases shows that the implemented cornering algorithm models the Prusa Mini firmware behavior well. The maximum discrepancy between the experiment and the simulation is $0.9 \%$ for jerk values of $15 \mathrm{\frac{mm}{s}}$.
+For the chosen case the layer duration is highly dependant on the set jerk values. For jerk values equal to the target printing velocity, the calculated time is expected to approach a constant velocity solution calculated analytically. Therefore, the acceleration and cornering algorithms have no influence on the print time of a layer. This case also represents the current state of the art, in which no further modeling is done. A comparison between state of the art and the modeled solution is also shown in \autoref{fig:acc_comp}. For jerk values close to zero, the printer is expected to slow almost to a full stop for each turn in the path. This result is similar to the simplest velocity trapezoid where entry and exit velocities are zero. The layer time for this edge case was also  validated by an analytical calculation. The comparison to experimental data for jerk values between these edge cases shows that the implemented cornering algorithm models the Prusa Mini firmware behavior well. The maximum discrepancy between the experiment and the simulation is $0.9 \%$ for jerk values of $15 \mathrm{\frac{mm}{s}}$.
 
 # Acknowledgements
 
