### 2024-06-13 cd7db44
remaining wording changes that had conflicts before
