+++ README.md
@@ -16,7 +16,7 @@ For the analysis of GCode, slicers or dedicated analyzer tools, such as [Prusa S
 
 ![comparison](https://media.githubusercontent.com/media/FAST-LB/pyGCodeDecode/main/paper/comparison.png)
 
-In detail, this package reads the target trajectory and commands for changing firmware settings from a GCode file. Subsequently it simulates a motion planner with acceleration and jerk / junction control. The more accurate modeling can only be achieved by replicating [grbl][grbl] and derivative firmwares specific movement planner solutions, such as Classic Jerk and Junction Deviation as an interpretation for Jerk. The simulation result describes the nozzle and extrusion axis position and velocity at every point in time and are easily accessible. This python package can be used to generate time dependent boundary conditions from a GCode file, needed in additive manufacturing simulations such as Fused Filament Fabrication. With implemented 3D plotting functions, it also can be useful as a GCode analyzer tool, to visualize local velocities to gain better process understanding.
+This package reads the target trajectory firmware settings changes from a GCode file. Subsequently, it simulates a motion planner with acceleration and jerk or junction control respectively. The more accurate modeling is achieved by replicating [grbl][grbl]'s and its derivatives' firmware-specific movement planner solutions, such as "Classic Jerk" and "Junction Deviation", as an interpretation for Jerk. The simulation result is a description of the nozzle and extrusion axis position and velocity over time and is easily accessible for further analysis. For example it can be used to generate time dependent boundary conditions, needed in additive manufacturing simulations for the Fused Filament Fabrication (FFF) process. The package includes 3D plotting functions, so it can be used to visualize local velocities before printing to improve your process understanding.
 
 
 ## Install pyGCodeDecode
