### 2024-06-13 #6465 @lukashof
> Some preliminary comments are as follows:
>
>     1. Please do add some key references in the main README.md file as well of your repository.
>
>     2. Also, request that any set of step-by instructions are presented in the README.md as an enumerated list and utilize appropriate grammar, punctuation and capitalizations in the instructions where necessary.
>
>     3. Although a statement of need is provided where it is mentioned the other software are inaccurate, it would be useful to see an image comparing them in the validation section. This would further strengthen the claim of the paper.
>
>     4. In addition to showing a comparison, the errors need to be tabulated as well.

Hello @RahulSundar and @zhubonan,

it took us a bit to get this done, but we finally addressed the comments.

Regarding 1. and 2.: We reworked the README. It should now introduce the package in a more concise way and be more readable overall.

Regarding 3.: The image comparing the state-of-the-art approach of simply displaying the target velocities from the GCode with our simulation is now also referenced in the validation section. It would be possible to move the image to that section entirely to distinguish the sections more clearly, but we felt, that it also adds value to the statement of need. We hope you agree.

Regarding 4.: The maximum error between our simulation and the experiment shown in the chart is now part of the text. We are already exceeding the length limit set by FOSS and therefore didn't add a table. The errors will also depend on the exact GCode and printer settings chosen for the test case. For that reason, we would like to avoid putting too much emphasis on the exact deviation from the experiment.

Thanks again for your invaluable feedback, @RahulSundar! We hope our changes meet your expectations. Let us know if you feel we missed something!
