+++ .github/workflows/paper.yaml
@@ -19,7 +19,7 @@ jobs:
           paper-path: paper/paper.md
 
       - name: upload PDF
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: ./paper/paper.pdf
