### 2024-06-13 9f1c993
version bump to 0.2.2
