### 2024-06-17 #6465 @zhubonan
@lukashof Thanks for the update!
> while working on the changes mentioned above, we realized that the page number in the footer seems to behave oddly in our compiled paper.

I think it is just a placeholder for the page number. Your paper has not been published, hence it does not have a page number. The PDF generated by the bot does have a page number (a fake one, probably) though.
Don't worry, we will run a set of final checks after acceptance.
