### 2024-07-05 #6465 @zhubonan
Hi @FelixFroelich  there are a few more steps before I can move the submission to the next stage. Please make a release of the package and deposit to a publicly available repository (for example, Zenodo) to get an DOI. This DOI will be linked to the JOSS publication. Please see the checklist above and make sure the metadata matches with that of the paper.

PS: Other recently published JOSS papers can act as examples on how the archive DOI is linked with the paper.
