### 2024-07-05 ebb6c72
Update keywords in CITATION.cff
