### 2024-07-08 d8dcabe
changes to citation.cff for final JOSS release
