### 2024-07-09 #6465 @lukashof
I just realized that Zenodo also creates a DOI not specific to any version. I understood that JOSS will use the one of the current version but in case you need it here is the overall DOI:
[10.5281/zenodo.12663658](https://zenodo.org/doi/10.5281/zenodo.12663658)
