### 2024-07-09 e8b7880
added DOI to README. JOSS DOI still tbd
