### 2024-07-12 #6465 @zhubonan
All look fine to me. There are a few flags of missing DOI but they are links to websites.
