### 2024-07-13 #6465 @kyleniemeyer
Hi @FelixFroelich, I have just a few paper edits for you before publishing:
- Can you try to break up the Summary section? It's currently a single very long paragraph.
- Please add commas after "e.g." throughout
- In the Methodology section, please combine the four citations using the `[@author1:2001; @author2:2001]` syntax
- The Altan and Eryildiz (2018) citation info looks incorrect; please replace with this BibTeX entry:
```bibtex
@article{Altan2018,
  title = {Effects of process parameters on the quality of PLA products fabricated by fused deposition modeling (FDM): surface roughness and tensile strength},
  volume = {60},
  ISSN = {0025-5300},
  DOI = {10.3139/120.111178},
  number = {5},
  journal = {Materials Testing},
  publisher = {Walter de Gruyter GmbH},
  author = {Altan,  Mirigul and Eryildiz,  Meltem and Gumus,  Beril and Kahraman,  Yusuf},
  year = {2018},
  month = may,
  pages = {471–477}
}
```
- It appears other journal article citations are missing volume/issue/page information as well (Gibson 2021, Luzanin 2019, Yadav 2023)
