+++ paper/paper.bib
@@ -29,39 +29,53 @@
   url = {https://www.klipper3d.org/Kinematics.html}
 }
 
-@book{Gibson:2021,
-  Author = {{Gibson}, I. and {Rosen}, D.},
+@book{Gibson2021,
+  address = {Cham, Switzerland},
+  author = {Gibson, Ian and Rosen, David and Stucker, Brent and Khorasani, Mahyar},
   doi = {10.1007/978-3-030-56127-7},
-  Booktitle = {Additive Manufacturing Technologies ~ISBN 978-3-030-56126-0},
-  Publisher = {Springer Nature Switzerland},
-  Title = {{Additive Manufacturing Technologies: Third Edition}},
-  Year = 2021,
-  page = {180}
+  edition = {3},
+  isbn = {978-3-030-56126-0},
+  pages = {171--200},
+  publisher = {Springer International Publishing},
+  title = {{Additive Manufacturing Technologies - Chapter 6: Material Extrusion}},
+  url = {https://link.springer.com/10.1007/978-3-030-56127-7},
+  year = {2021}
 }
 
-
-@article{Yadav:2023,
+@article{Yadav2023,
+  author = {Yadav, Krishna and Rohilla, Shivam and Ali, Arshad and Yadav, Mohit and Chhabra, Deepak},
   doi = {10.1007/s11665-023-08476-2},
-  Author = {{Yadav}, K. and {Rohilla}, S and {Ali}, A},
-  Journal = {Journal of Materials Engineering and Performance},
-  Title = {{Effect of Speed, Acceleration, and Jerk on Surface Roughness of FDM-Fabricated Parts}},
-  Year = 2023
+  issn = {1544-1024},
+  journal = {Journal of Materials Engineering and Performance},
+  title = {{Effect of Speed, Acceleration, and Jerk on Surface Roughness of FDM-Fabricated Parts}},
+  url = {https://doi.org/10.1007/s11665-023-08476-2},
+  year = {2023}
 }
 
-@article{Luzanin:2019,
+@article{Luzanin2019,
+  author = {Luzanin, Ognjan and Movrin, Dejan and Stathopoulos, Vassilis and Pandis, Pavlos and Radusin, Tanja and Guduric, Vera},
   doi = {10.1108/RPJ-12-2018-0316},
-  Author = {{Luzanin}, O. and {Movrin}, D.},
-  Journal = {Rapid Prototyping},
-  Title = {{Impact of processing parameters on tensile strength, in-process crystallinity and mesostructure in FDM-fabricated PLA specimens}},
-  Year = 2019
+  issn = {13552546},
+  journal = {Rapid Prototyping Journal},
+  number = {8},
+  pages = {1398--1410},
+  title = {{Impact of processing parameters on tensile strength, in-process crystallinity and mesostructure in FDM-fabricated PLA specimens}},
+  volume = {25},
+  year = {2019}
 }
 
-@article{Altan:2018,
-  doi = {10.3139/120.111178},
-  Author = {{Altan}, M. and {Eryildiz}, M.},
-  Journal = {MATERIALS TESTING FOR JOINING AND ADDITIVE MANUFACTURING APPLICATIONS},
-  Title = {{Effects of process parameters on the quality of PLA products fabricated by fused deposition modeling (FDM): surface roughness and tensile strength}},
-  Year = 2018
+@article{Altan2018,
+  title = {Effects of process parameters on the quality of PLA products fabricated by fused deposition modeling (FDM): surface roughness and tensile strength},
+  volume = {60},
+  ISSN = {0025-5300},
+  DOI = {10.3139/120.111178},
+  number = {5},
+  journal = {Materials Testing},
[…]
