+++ paper/paper.md
@@ -48,7 +48,9 @@ bibliography: paper.bib
 # Summary
 
 The Machine instructions for material extrusion processes (MEX), such as the fused filament fabrication (FFF) process, are typically provided as GCode, which can be generated by a variety of slicer programs. The 3D model of the part is sliced into multiple layers and a tool path is created for each according to the parameters for infill, perimeters supports and other structures [@Gibson:2021]. The exported GCode consists of a list of commands specifying target points in space for the tool as well as the amount of material to be extruded. Additionally, process parameters such as temperatures, velocities or cooling fan speeds are set and changed during printing according to the GCode.
+
 However, the GCode itself does not accurately reflect the eventual printing process. It is interpreted by the printer's firmware that plans the trajectory taking into account the machine's limitations. In particular, the specified maximum printing speed, acceleration and jerk have an influence on the resulting path velocities. These influence both the mechanical properties such as the resulting crystallinity when processing semi-crystalline thermoplastics [@Luzanin:2019] and the tensile strength or surface roughness [@Altan:2018]. The direct influence of firmware parameters such as "jerk settings" and acceleration on surface roughness was also shown in [@Yadav:2023]. This means that print results and print times for the same GCode path can vary when using different printers, even if many printers use similar firmware.
+
 Setting a higher target printing velocity on a machine with insufficient acceleration capabilities will lead to a large difference between target and actual printing velocity as illustrated in \autoref{fig:acc_comp}. This can lead to unexpected behavior and a slower print than anticipated. Many slicers will predict the progression of the print but these predictions might deviate significantly from the actual process. A good understanding and accurate modeling of trajectory behaviors can contribute significantly to the improvement of slicing algorithms and printer hardware through the virtual evaluation of GCode. In addition, modeling of those behaviors enables more accurate virtual replication of the process through process simulations such as thermomechanical modeling and small-scale fluid simulations.
 <span style="font-variant:small-caps;">pyGCodeDecode</span> is a Python package for GCode interpretation and MEX Firmware simulation. The package was developed to enable researchers and users to better understand time-dependent process variables and enable a more accurate study of the printing process.
 
@@ -57,11 +59,12 @@ Setting a higher target printing velocity on a machine with insufficient acceler
 # Statement of need
 
 There are several software tools to visualize GCode file data. For example, in various slicer programs such as <span style="font-variant:small-caps;">Prusa Slicer</span> [@prusa_slicer] or <span style="font-variant:small-caps;">Cura</span> [@cura], but also in web applications and printer control applications such as <span style="font-variant:small-caps;">Octoprint</span> [@octoprint], <span style="font-variant:small-caps;">Repetier-Host</span> [@repetier_host], <span style="font-variant:small-caps;">NC Viewer</span> [@ncviewer] or <span style="font-variant:small-caps;">gCode Viewer</span> [@gcodeviewer]. These tools can read the position of the GCode coordinates and interpolate between the points to create motion paths. It is possible to distinguish between printing and traversing motions to preview the part. The additional information in the GCode, such as target print speed or temperature, can also be displayed in most cases. However, currently available tools are unable to accurately simulate the behavior of the printer, including acceleration and deceleration. This can lead to inaccurate time predictions and potentially undetected deviations from expected process conditions.
+
[…]
