+++ paper/paper.bib
@@ -29,7 +29,7 @@
   url = {https://www.klipper3d.org/Kinematics.html}
 }
 
-@book{Gibson2021,
+@book{Gibson:2021,
   address = {Cham, Switzerland},
   author = {Gibson, Ian and Rosen, David and Stucker, Brent and Khorasani, Mahyar},
   doi = {10.1007/978-3-030-56127-7},
@@ -42,7 +42,7 @@
   year = {2021}
 }
 
-@article{Yadav2023,
+@article{Yadav:2023,
   author = {Yadav, Krishna and Rohilla, Shivam and Ali, Arshad and Yadav, Mohit and Chhabra, Deepak},
   doi = {10.1007/s11665-023-08476-2},
   issn = {1544-1024},
@@ -52,7 +52,7 @@
   year = {2023}
 }
 
-@article{Luzanin2019,
+@article{Luzanin:2019,
   author = {Luzanin, Ognjan and Movrin, Dejan and Stathopoulos, Vassilis and Pandis, Pavlos and Radusin, Tanja and Guduric, Vera},
   doi = {10.1108/RPJ-12-2018-0316},
   issn = {13552546},
@@ -64,7 +64,7 @@
   year = {2019}
 }
 
-@article{Altan2018,
+@article{Altan:2018,
   title = {Effects of process parameters on the quality of PLA products fabricated by fused deposition modeling (FDM): surface roughness and tensile strength},
   volume = {60},
   ISSN = {0025-5300},
