### 2024-07-15 #6465 @FelixFroelich
Hi @kyleniemeyer, thank you for your comments. We have implemented them as best we could. We still have a few questions/comments:

- Gibson2021 refers to a book. We have now added the chapter, the corresponding page numbers, and the edition.
- Unfortunately, we could not find a volume and page numbers for Yadav2023. We were also unable to find the publication in the corresponding volumes of the journal (July and August 2023).
- No year is given for the references that refer to software. Here, for example, it says Xander Luciano, n.d.. Is that a problem? We could also cite a software version and give the corresponding year of publication. However, this is not possible for all software.

Thank you very much for your support!
