### 2024-07-15 #6465 @kyleniemeyer
@FelixFroelich thanks, those changes look good. For the software citations, if you aren't citing a specific version with a release date, one option is to include the year of the latest release/change (which may just be 2024 for something that is still actively under development).
