### 2024-07-15 abd37a6
Update paper.bib
