+++ paper/paper.bib
@@ -9,6 +9,7 @@
 @software{marlin,
   author = {Scott Lahteine},
   title = {Marlin Firmware},
+  year = {2024},
   version = {2.1.2.2},
   url = {https://github.com/MarlinFirmware/Marlin}
 }
@@ -17,6 +18,7 @@
   author = {Scott Lahteine},
   title = {Marlin Documentation},
   urldate = {2024-04-09},
+  year = {2024},
   journal = {},
   url = {https://marlinfw.org/}
 }
@@ -25,6 +27,7 @@
   author = {Klipper3d},
   title = {Klipper Documentation},
   urldate = {2024-04-09},
+  year = {2023},
   journal = {},
   url = {https://www.klipper3d.org/Kinematics.html}
 }
@@ -111,6 +114,7 @@
   title = {{NC Viewer}},
   url = {https://ncviewer.com/},
   urldate = {2024-04-09},
+  year = {2018},
   version = {1.1.3}
 }
 
@@ -118,6 +122,7 @@
   author = {{Alex Ustyantsev}},
   title = {{gCodeViewer}},
   urldate = {2024-04-09},
+  year = {2020},
   url = {https://gcode.ws/}
 }
 
@@ -125,5 +130,6 @@
   author = {{Hot-World GmbH & Co. KG}},
   title = {{Repetier-Host}},
   url = {https://www.repetier.com/},
+  year = {2024},
   version = {2.3.2}
 }
