### 2024-08-17 #6465 @RahulSundar
Dear @editorialbot @kyleniemeyer , Is there a way to get acknowledged for the JOSS review efforts on ORCID?
