### 2024-08-19 #6465 @kyleniemeyer
@RahulSundar unfortunately no, we don't have a formal way of doing this right now. However, your effort as a reviewer *is* acknowledged on the [paper's web page](https://joss.theoj.org/papers/10.21105/joss.06465), and you also have a reviewer profile page that shows all your reviews: https://joss.theoj.org/papers/reviewed_by/@RahulSundar
